requirements.txt
flask
python-dotenv
requests
bs4